Move screen datatypes and add tests for instructions
