+++ src/main/java/ca/craigthomas/yacoco3e/datatypes/screen/G6RScreenMode.java
@@ -2,11 +2,13 @@
  * Copyright (C) 2018 Craig Thomas
  * This project uses an MIT style license - see LICENSE for details.
  */
-package ca.craigthomas.yacoco3e.datatypes;
+package ca.craigthomas.yacoco3e.datatypes.screen;
+
+import ca.craigthomas.yacoco3e.datatypes.UnsignedByte;
 
 import java.awt.*;
 
-public class G6CScreenMode extends ScreenMode
+public class G6RScreenMode extends ScreenMode
 {
     /* Screen size for the mode */
     private static final int SCREEN_WIDTH = 320;
@@ -15,29 +17,31 @@ public class G6CScreenMode extends ScreenMode
     /* Block definitions */
     private static final int BLOCK_WIDTH = 2;
     private static final int BLOCK_HEIGHT = 1;
-    private static final int BLOCKS_PER_BYTE = 4;
+    private static final int BLOCKS_PER_BYTE = 8;
 
-    /* Color definitions for graphics G6C mode */
+    /* Color definitions for graphics G3R mode */
     private final Color colors[][] = {
-        {
-            // Color Mode 0
-            new Color(40, 224, 40, 255),   /* Green */
-            new Color(240, 240, 112, 255), /* Yellow */
-            new Color(32, 32, 216, 255),   /* Blue */
-            new Color(168, 32, 32, 255),   /* Red */
-        }, {
-            // Color Mode 1
-            new Color(240, 240, 240, 255), /* White */
-            new Color(40, 168, 168, 255),  /* Cyan */
-            new Color(211, 97, 250, 255),  /* Magenta */
-            new Color(240, 136, 40, 255),  /* Orange */
-        }
+            {
+                // Color Mode 0
+                new Color(0, 0, 0, 255),        /* Black */
+                new Color(40, 224, 40, 255),    /* Green */
+                new Color(240, 136, 40, 255),   /* Orange Artifact */
+                new Color(32, 32, 216, 255),    /* Blue Artifact */
+            }, {
+                // Color Mode 1
+                new Color(0, 0, 0, 255),        /* Black */
+                new Color(32, 32, 216, 255),    /* Blue Artifact */
+                new Color(255, 60, 32, 255),   /* Orange Artifact */
+                new Color(240, 240, 240, 255),  /* White */
+            }
     };
+
+    private final Color background = new Color(240, 240, 240, 255);
 
     // The color mode to apply
     private int colorMode;
 
-    public G6CScreenMode(int scale, int colorMode) {
+    public G6RScreenMode(int scale, int colorMode) {
         this.scale = scale;
         this.width = SCREEN_WIDTH;
         this.height = SCREEN_HEIGHT;
@@ -48,13 +52,13 @@ public class G6CScreenMode extends ScreenMode
     @Override
     public void refreshScreen() {
         Graphics2D graphics = backBuffer.createGraphics();
-        graphics.setColor(colors[colorMode][0]);
+        graphics.setColor(background);
         graphics.fillRect(0, 0, width * scale, height * scale);
 
         int memoryPointer = memoryOffset;
 
         for (int y = 0; y < 192; y++) {
-            for (int x = 0; x < (128 / BLOCKS_PER_BYTE); x++) {
+            for (int x = 0; x < 32; x++) {
                 UnsignedByte value = io.readPhysicalByte(memoryPointer);
                 drawCharacter(value, x, y);
                 memoryPointer++;
@@ -74,7 +78,7 @@ public class G6CScreenMode extends ScreenMode
 
     private void drawCharacter(UnsignedByte value, int col, int row) {
         /* Translated position in pixels */
-        int x = 32 + (col * (BLOCK_WIDTH * BLOCKS_PER_BYTE));
+        int x = 32 + (col * 8);
         int y = 24 + (row * BLOCK_HEIGHT);
 
         /* Pixel 1 */
