+++ src/main/java/ca/craigthomas/yacoco3e/datatypes/screen/SG24ScreenMode.java
@@ -2,25 +2,27 @@
  * Copyright (C) 2018 Craig Thomas
  * This project uses an MIT style license - see LICENSE for details.
  */
-package ca.craigthomas.yacoco3e.datatypes;
+package ca.craigthomas.yacoco3e.datatypes.screen;
+
+import ca.craigthomas.yacoco3e.datatypes.UnsignedByte;
 
 import java.awt.*;
 
-public class SG8ScreenMode extends ScreenMode
+public class SG24ScreenMode extends ScreenMode
 {
     /* Semi-graphics color constants */
     private static final int GREEN = 0;
     private static final int BLACK = 8;
 
-    /* Screen size for semi-graphics 4 mode */
+    /* Screen size for semi-graphics 24 mode */
     private static final int SCREEN_WIDTH = 320;
     private static final int SCREEN_HEIGHT = 240;
 
     /* Block definitions */
     private static final int BLOCK_WIDTH = 4;
-    private static final int BLOCK_HEIGHT = 3;
+    private static final int BLOCK_HEIGHT = 1;
 
-    /* Color definitions for semi-graphics 8 mode */
+    /* Color definitions for semi-graphics 24 mode */
     private final Color colors[] = {
             new Color(40, 224, 40, 255),   /* Green */
             new Color(240, 240, 112, 255), /* Yellow */
@@ -38,7 +40,7 @@ public class SG8ScreenMode extends ScreenMode
     // The background color
     private int backColor;
 
-    public SG8ScreenMode(int scale) {
+    public SG24ScreenMode(int scale) {
         this.scale = scale;
         this.width = SCREEN_WIDTH;
         this.height = SCREEN_HEIGHT;
@@ -63,10 +65,10 @@ public class SG8ScreenMode extends ScreenMode
 
         int memoryPointer = memoryOffset;
 
-        for (int y = 0; y < 64; y++) {
+        for (int y = 0; y < 256; y++) {
             for (int x = 0; x < 32; x++) {
                 UnsignedByte value = io.readPhysicalByte(memoryPointer);
-                drawCharacter(value, x, y, y % 4);
+                drawCharacter(value, x, y, y % 12);
                 memoryPointer++;
             }
         }
@@ -75,7 +77,7 @@ public class SG8ScreenMode extends ScreenMode
     }
 
     /**
-     * Draws an SG4 block. Blocks are of width by height in size and
+     * Draws an SG24 block. Blocks are of width by height in size and
      * have a foreground color.
      *
      * @param col the column to draw the block in
@@ -91,9 +93,9 @@ public class SG8ScreenMode extends ScreenMode
     }
 
     /**
-     * Draw a Semi Graphics 6 character on the screen at the specified
-     * column and row. Each character is 8 x 6 in size. The character
-     * cell is further broken down into 2 subcells of 4 x 3 blocks each
+     * Draw a Semi Graphics 24 character on the screen at the specified
+     * column and row. Each character is 4 x 1 in size. The character
+     * cell is further broken down into 2 subcells of 2 x 1 blocks each
      * in the following configuration:
      *
      *    +----+----+
@@ -107,19 +109,19 @@ public class SG8ScreenMode extends ScreenMode
     private void drawCharacter(UnsignedByte value, int col, int row, int rowMod) {
         /* Translated position in pixels */
         int x = 32 + (col * 8);
-        int y = 24 + (row * 3);
+        int y = 24 + row;
 
         /* Background colors */
         int back = BLACK;
         int color = (value.getShort() & 0x70) >> 4;
 
         /* Subcell B */
-        int mask = rowMod == 0 || rowMod == 1 ? 0x8 : 0x2;
+        int mask = rowMod == 0 || rowMod == 1 || rowMod == 2 || rowMod == 3 || rowMod == 4 || rowMod == 5 || rowMod == 6 ? 0x8 : 0x2;
         int on = value.isMasked(mask) ? 1 : 0;
         drawBlock(x, y, on == 1 ? color : back);
 
         /* Subcell B */
-        mask = rowMod == 0 || rowMod == 1 ? 0x4 : 0x1;
+        mask = rowMod == 0 || rowMod == 1 || rowMod == 2 || rowMod == 3 || rowMod == 4 || rowMod == 5 || rowMod == 6 ? 0x4 : 0x1;
         on = value.isMasked(mask) ? 1 : 0;
         drawBlock(x + BLOCK_WIDTH, y, on == 1 ? color : back);
     }
