+++ src/main/java/ca/craigthomas/yacoco3e/datatypes/screen/SG6ScreenMode.java
@@ -2,11 +2,13 @@
  * Copyright (C) 2018 Craig Thomas
  * This project uses an MIT style license - see LICENSE for details.
  */
-package ca.craigthomas.yacoco3e.datatypes;
+package ca.craigthomas.yacoco3e.datatypes.screen;
+
+import ca.craigthomas.yacoco3e.datatypes.UnsignedByte;
 
 import java.awt.*;
 
-public class SG12ScreenMode extends ScreenMode
+public class SG6ScreenMode extends ScreenMode
 {
     /* Semi-graphics color constants */
     private static final int GREEN = 0;
@@ -18,9 +20,9 @@ public class SG12ScreenMode extends ScreenMode
 
     /* Block definitions */
     private static final int BLOCK_WIDTH = 4;
-    private static final int BLOCK_HEIGHT = 2;
+    private static final int BLOCK_HEIGHT = 4;
 
-    /* Color definitions for semi-graphics 12 mode */
+    /* Color definitions for semi-graphics 4 mode */
     private final Color colors[] = {
             new Color(40, 224, 40, 255),   /* Green */
             new Color(240, 240, 112, 255), /* Yellow */
@@ -37,11 +39,14 @@ public class SG12ScreenMode extends ScreenMode
     private int foreColor;
     // The background color
     private int backColor;
+    // The color set to use
+    private int colorSet;
 
-    public SG12ScreenMode(int scale) {
+    public SG6ScreenMode(int scale, int colorSet) {
         this.scale = scale;
         this.width = SCREEN_WIDTH;
         this.height = SCREEN_HEIGHT;
+        this.colorSet = colorSet;
         foreColor = GREEN;
         backColor = BLACK;
         createBackBuffer();
@@ -62,8 +67,7 @@ public class SG12ScreenMode extends ScreenMode
         );
 
         int memoryPointer = memoryOffset;
-
-        for (int y = 0; y < 96; y++) {
+        for (int y = 0; y < 16; y++) {
             for (int x = 0; x < 32; x++) {
                 UnsignedByte value = io.readPhysicalByte(memoryPointer);
                 drawCharacter(value, x, y);
@@ -75,7 +79,7 @@ public class SG12ScreenMode extends ScreenMode
     }
 
     /**
-     * Draws an SG4 block. Blocks are of width by height in size and
+     * Draws an SG6 block. Blocks are of width by height in size and
      * have a foreground color.
      *
      * @param col the column to draw the block in
@@ -91,13 +95,17 @@ public class SG12ScreenMode extends ScreenMode
     }
 
     /**
-     * Draw a Semi Graphics 12 character on the screen at the specified
-     * column and row. Each character is 8 x 2 in size. The character
-     * cell is further broken down into 2 subcells of 4 x 2 blocks each
+     * Draw a Semi Graphics 6 character on the screen at the specified
+     * column and row. Each character is 8 x 12 in size. The character
+     * cell is further broken down into 6 subcells of 4 x 4 blocks each
      * in the following configuration:
      *
      *    +----+----+
      *    |  A |  B |
+     *    +----+----+
+     *    |  C |  D |
+     *    +----+----+
+     *    |  E |  F |
      *    +----+----+
      *
      * @param value the value of the byte to write
@@ -106,19 +114,39 @@ public class SG12ScreenMode extends ScreenMode
      */
     private void drawCharacter(UnsignedByte value, int col, int row) {
         /* Translated position in pixels */
-        int x = 32 + (col * (BLOCK_WIDTH * 2));
-        int y = 24 + (row * BLOCK_HEIGHT);
+        int x = 32 + (col * 8);
+        int y = 24 + (row * 12);
 
         /* Background colors */
-        int color = (value.getShort() & 0x70) >> 4;
-        if (!value.isMasked(0x80)) {
-            color = BLACK;
+        int back = BLACK;
+
+        int color = (value.getShort() & 0xC0) >> 6;
+        if (colorSet == 1) {
+            color = color + 4;
         }
 
-        /* Subcell B */
-        drawBlock(x, y, value.isMasked(0x8) ? color : BLACK);
+        /* Subcell A */
+        int on = value.isMasked(0x20) ? 1 : 0;
+        drawBlock(x, y, on == 1 ? color : back);
 
         /* Subcell B */
-        drawBlock(x + BLOCK_WIDTH, y, value.isMasked(0x4) ? color : BLACK);
+        on = value.isMasked(0x10) ? 1 : 0;
+        drawBlock(x + BLOCK_WIDTH, y, on == 1 ? color : back);
+
+        /* Subcell C */
+        on = value.isMasked(0x8) ? 1 : 0;
+        drawBlock(x, y + BLOCK_HEIGHT, on == 1 ? color : back);
+
+        /* Subcell D */
+        on = value.isMasked(0x4) ? 1 : 0;
+        drawBlock(x + BLOCK_WIDTH, y + BLOCK_HEIGHT, on == 1 ? color : back);
+
+        /* Subcell E */
+        on = value.isMasked(0x2) ? 1 : 0;
+        drawBlock(x, y + (BLOCK_HEIGHT * 2), on == 1 ? color : back);
+
+        /* Subcell F */
+        on = value.isMasked(0x1) ? 1 : 0;
+        drawBlock(x + BLOCK_WIDTH, y + (BLOCK_HEIGHT * 2), on == 1 ? color : back);
     }
 }
