+++ src/main/java/ca/craigthomas/yacoco3e/datatypes/screen/ScreenMode.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2017-2018 Craig Thomas
  * This project uses an MIT style license - see LICENSE for details.
  */
-package ca.craigthomas.yacoco3e.datatypes;
+package ca.craigthomas.yacoco3e.datatypes.screen;
 
 import ca.craigthomas.yacoco3e.components.IOController;
